requirements.txt
openai>=1.0.0
python-dotenv>=1.0.0
pyaudio>=0.2.11
wave
plotly>=5.0.0
scikit-learn>=1.0.0
sentence-transformers>=2.0.0
together>=0.2.0
numpy>=1.21.0
beautifulsoup4>=4.9.0
requests>=2.25.0
